requirements.txt
elasticsearch[async]
mtcnn-onnxruntime
imutils
tqdm
torch
torchvision
numpy
pandas
